tests/wpt/tests/web-animations/interfaces/Animation/id.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation.id</title>
<link rel="help" href="https://drafts.csswg.org/web-animations/#dom-animation-id">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

test(t => {
  const div = createDiv(t);
  const animation = div.animate({}, 100 * MS_PER_SEC);
  assert_equals(animation.id, '', 'id for Animation is initially empty');
}, 'Animation.id initial value');

test(t => {
  const div = createDiv(t);
  const animation = div.animate({}, 100 * MS_PER_SEC);
  animation.id = 'anim';

  assert_equals(animation.id, 'anim', 'animation.id reflects the value set');
}, 'Animation.id setter');

</script>
</body>
